# src/config/swagger.yaml
openapi: 3.0.0
info:
  title: Buildup Investment
  description: API for managing company names
  version: 0.1.9
servers:
  - url: http://localhost:3000/api
    description: Local server
paths:
...
  /create-apartments:
    post:
      summary: Generate apartments based on a floor plan ID and building ID
      operationId: createApartments
      tags:
        - Apartments
      requestBody:
        description: The floor plan ID, building ID, and name used to generate apartments.
        content:
          application/json:
            required: true
            schema:
              type: object
              properties:
                floor_plan_id:
                  type: integer
                  description: The ID of the floor plan to generate apartments for.
                  example: 1
                building_id:
                  type: integer
                  description: The ID of the building for the apartments.
                  example: 1
                name:
                  type: string
                  description: The name or label for the floor plan.
                  example: "Plan A"
              required:
                - floor_plan_id
                - building_id
                - name
      responses:
        "201":
          description: Apartments generated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: SUCCESS
                  message:
                    type: string
                    example: Apartments generated successfully.
                  apartments:
                    type: array
                    items:
                      type: object
                      properties:
                        floor:
                          type: integer
                          example: 1
                        apartments:
                          type: array
                          items:
                            type: object
                            properties:
                              flat_id:
                                type: integer
                                example: 1
                              flat_number:
                                type: integer
                                example: 101
                              building_id:
                                type: integer
                                example: 1
                              floor_plan_id:
                                type: integer
                                example: 1
                              status:
                                type: string
                                enum: ["free", "reserved", "sold"]
                                default: "free"
                                example: "free"
                              image:
                                type: string
                                description: URL or path to apartment image
                                nullable: true
                                example: null
                              square_meters:
                                type: number
                                format: float
                                description: Apartment area in square meters
                                default: 0
                                example: 0
        "400":
          description: Invalid input or duplicate apartments.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ERROR
                  error:
                    type: string
                    enum:
                      [
                        BUILDING_NOT_FOUND,
                        FLOOR_PLAN_NOT_FOUND,
                        DUPLICATE_APARTMENTS,
                      ]
                    example: BUILDING_NOT_FOUND
                  message:
                    type: string
                    example: "Building with ID 1 does not exist."
        "500":
          description: An error occurred while generating apartments.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ERROR
                  error:
                    type: string
                    example: SERVER_ERROR
                  message:
                    type: string
                    example: An error occurred while generating apartments.
                  details:
                    type: string
                    example: Some detailed error message (only in development mode)